spider_web_backend/requirements.txt
asgiref==3.4.1
Django==3.2.20
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.6
pyparsing==3.1.0
pytz==2023.3
PyYAML==6.0.1
typing_extensions==4.1.1
uritemplate==4.1.1

elasticsearch~=8.9.0
pymongo~=3.12.1